rename strings.ts to stringTypes.ts
