+++ src/controllers/connection/index.ts
@@ -3,7 +3,7 @@ import { pino } from 'pino'
 import { Body, Get, Path, Post, Produces, Query, Request, Route, Security, SuccessResponse } from 'tsoa'
 import { injectable } from 'tsyringe'
 
-import { pinCodeRegex, type PIN_CODE, type UUID } from '../../models/strings.js'
+import { pinCodeRegex, type PIN_CODE, type UUID } from '../../models/stringTypes.js'
 import ConnectionTemplates from '../../views/connection/connection.js'
 
 import { DatabaseTimeoutError, InternalError, InvalidInputError, NotFoundError } from '../../errors.js'
