+++ src/controllers/queries/__tests__/helpers.ts
@@ -3,7 +3,7 @@ import sinon from 'sinon'
 
 import Database from '../../../models/db/index.js'
 import { ConnectionRow, QueryRow } from '../../../models/db/types.js'
-import { UUID } from '../../../models/strings.js'
+import { UUID } from '../../../models/stringTypes.js'
 import VeritableCloudagent from '../../../models/veritableCloudagent/index.js'
 import QueriesTemplates from '../../../views/queries/queries.js'
 import QueryListTemplates from '../../../views/queries/queriesList.js'
