+++ src/models/arrays.ts
@@ -1,5 +1,5 @@
 import { UInt } from './numbers.js'
-import { UUID } from './strings.js'
+import { UUID } from './stringTypes.js'
 
 export type PartialQuery = { connectionId: UUID; productId: string; quantity: number }
 export type PartialQueryPayload = {
