+++ src/models/stringTypes.ts
