+++ src/server.ts
@@ -11,7 +11,7 @@ import { ValidateError } from 'tsoa'
 import { Env } from './env/index.js'
 import { ForbiddenError, HttpError } from './errors.js'
 import { ILogger, Logger } from './logger.js'
-import { UUID } from './models/strings.js'
+import { UUID } from './models/stringTypes.js'
 import { RegisterRoutes } from './routes.js'
 import ConnectionEvents from './services/connectionEvents.js'
 import CredentialEvents from './services/credentialEvents/index.js'
