+++ test/helpers/connection.ts
@@ -7,7 +7,7 @@ import sinon from 'sinon'
 import Database from '../../src/models/db/index.js'
 import { ConnectionRow } from '../../src/models/db/types.js'
 import EmailService from '../../src/models/emailService/index.js'
-import { CountryCode, UUID } from '../../src/models/strings.js'
+import { CountryCode, UUID } from '../../src/models/stringTypes.js'
 import VeritableCloudagent from '../../src/models/veritableCloudagent/index.js'
 import createHttpServer from '../../src/server.js'
 import VeritableCloudagentEvents from '../../src/services/veritableCloudagentEvents.js'
