+++ test/integration/newConnection.test.ts
@@ -2,7 +2,7 @@ import * as chai from 'chai'
 import chaiAsPromised from 'chai-as-promised'
 import { afterEach, beforeEach, describe, it } from 'mocha'
 import sinon from 'sinon'
-import { CountryCode } from '../../src/models/strings.js'
+import { CountryCode } from '../../src/models/stringTypes.js'
 import { cleanupCloudagent, cleanupDatabase } from '../helpers/cleanup.js'
 import { setupTwoPartyContext, TwoPartyContext } from '../helpers/connection.js'
 import { alice, socrataCompany } from '../helpers/fixtures.js'
